[Win] Sort groups
We might have to resort to doing this custom somewhere. The sorting
mechanism isn't extensible and we probably want the Misc group at the
bottom like VS does rather than purely by alpha sort.

The built in sort is a bit brain-dead in that it does it _before_
applying the value converter, which means my last trick doesn't sort
misc to the bottom.