MockObjectEditor and some docs
It may be neccessary to re-query values sometimes and it's too difficult
to setup a Mock<> to do this, so we'll create a simple mock class that
accepts arbitrary properties.

Why do this instead of just using a reflection editor? A lot of times it
becomes neccessary to control what the IPropertyInfo returns to test the
various scenarios. In some cases (resources, bindings) we can't do this
via a reflection editor.