[Core] Fix editor requests and optimize
Since IObjectEditors were not equatable, we were adding duplicate
editors every time the selected items list changed. Instead of
implementing equality on object editors (which we may still do for other
reasons) we'll request editors one at a time. This messes with the
current distinction between ObjectViewModel and PanelViewModel, but it's
largely pointless. This way we avoid recreating editors (which might be
expensive) altogether.

We also add a .Target property to the object for identity. This avoids
having to build a map between selected objects to their editors. The
selected items list itself should never be terribly long, so doing a
linear comparison is no big deal.