[Mac] Mostly work around tabbing issue
There seems to be an issue with the way the key view loop is calculated
with the NSTableView. This patch fixes it 95%, but I can still cause
it to break if I scroll the table really quickly and then go back to
where I was and attempt to keep tabbing forward/backward.

This workaround overrides a private selector, as suggested by:
https://forums.developer.apple.com/thread/16528#184148