[Win] Swap itemssource on arrange
Rather than trying to keep multiple views of the same collection up to
date, the panel view model stuffs name-arranged properties into the
first element of the grouped.

Unfortunately since binding to it is treated as a property, we have to
force the binding when that changes. Here we're doing it in the most
simple fashion, but we can do it more accurately if it causes
performance issues.