[Mac] Refactor & Cleanup editors
The update-from-model method was named in a way that sounded like it did the opposite,
so it's been clarified. This also adds a method for handling when the view model has
changed as this is distinct from needing to update the values. Populating combo boxes
from the VM for example could happen too often because it was tied to updating values.
Additionally this removes the need to duplicate the value check and update every subclass.

There's more cleanup to do on errors, but that requires some more work in the VMs first.