[Core] Fix AsyncWorkQueue
This fixes an issue (demonstrated in included test) where a
child/sibling to another requester is added when that requester is not
the active. The logic that was present in the reqest to set related
requester workers to active was not present when the queue was advanced.

This also significantly reduces the work the request work method does by
simply tracking the active requester. The original comment about small
queues turned out to be false and it is likely to see queues as long as
2x the number of properties.