[Tests] Simplify mocks storage
Instead of having a property bag instead of MockObjectEditor and
MockPropertyInfo which both get used for the tests standalone app, we
really only need the one in the Editor. Prompting this comes from the
fact that we need to test more than just Local values, and
MockObjectEditor already has a basis for supporting this, while the
MockPropertyInfo and MockControl things don't.

The trick here is that MockPropertyInfo needs some control over the
values for type conversions (instead of just having a giant switch in
the editor). To solve this, the property infos lose their set overrides
and gain a conversion method.