[Core] Add SimpleCollectionView
This still needs more tests and has plenty of room for optimization, but
the existing bespoke implementations are no more optimized. A central
place to add tests and optimizations is an obvious win.

Bringing in the WPF style collection view from Moonlight was considered,
but ultimately it comes with a lot of baggage and would need to be
adapted to our needs.