[Tests] Remove invalid part of test
Technically because the VM will set the new value as a local it is not
up to the VM whether the value change event fires, it is ultimately up
to the editor. If the editor determines that the value is still unset it
can ignore the source change which wouldn't fire, otherwise it would.