Added support for theming proppy via VS color theme
This code is packaged with proppy source, so we can keep the
color list up to date here, but it's intended to be compiled with
the client app (which is straightforward - the VS Designer code just
points to this source file in the proppy git submodule).

Everything seems to work well here, except that there are
some brush names that I couldn't find matches for in the XET
source. Those brushes are commented out in AddBrushes. Eric -
for those brushes, the ones commented out, can you let me know
where you got those brush names / colors from. I'll then update
the mapping to handle them too.