[Win] Fix NRE in ChoiceControl on unload
In VS, unloading controls seems to want to unset their styles and
therefore their templates which triggers updates not seen in the
standalone app.

This fixes the container generator status change when the template is
changed to null. When it's null we don't care about the contents so
we'll just return.