[Win] Move to preview template selector
This moves the "no preview" fallback to a template in the selector. The
problem before was the contentpresenter would always fallback to a
ToString if it didn't have a template for the datatype, leaving no way
to control visibility of No Preview for previews that weren't a solid
cover (like a string instead of a color).

This also fixes placement of string previews (fixes #238)