[Core] Trigger Value update after individual choices up to date
While the most correct thing is for the UI to reflect the values of the
individual choices when they are changed, a reasonable usage is to wait
for the value to update and then trigger the update. Reordering Value to
update after the flags ensure that they are up to date when Value is
triggered through INPC.

Given that the flags are really the source of truth for the values in
the combinable property, it makes sense for them to be first.