[Win] Add properties host flag
We need a parent control to pull resources from to pass them through to
child windows. Originally we were searching for the panel, but the panel
includes all the search/arrange features, we may just need the list of
properties. So now we have a marker to hint at which controls would
contain the resources.