[Core] Rework CollectionPropertyViewModel
In order to track rows for display purposes and in order to provide a
clean way to get at the "real" name of the child, we've moved to using a
VM per child. Since the main way to get a real name of an editable
target is to request an object editor, we introduce a cached provider so
the PanelViewModel doesn't have to get its own copy when the targets are
selected.

This collections tend to be small, but if we run into issues with the
amount of things we're keeping alive we can easily move to on-demand
requests.