[Core] TypeName -> TargetType
This moves away from a straight name on the IObjectEditor to a ITypeInfo
describing the type. As this is currently expect to be non-null, it was
added to the tests (as well as some missing .Target assignments).

Caught up in this change were some tests being moved down a rank from
PanelViewModelTests to PropertiesViewModelTests.