[Core] Split SourceDescriptor out from ValueDescriptor
ValueDescriptor ended up being used for overlapping purposes:
binding/resources for the source but then for some values like objects
and predfined values it had extra information not contained in the
Value. SourceDescriptor was added to hold things like bindings or
resources that describe where the value came from, while ValueDescriptor
will continue to hold extra details about the value.

There were two possible approaches to this, to add the object property
as-is in this commit, or to add bespoke properties for a Resource and
Binding. Ultimately both approaches have fairly similar level pros and
cons. What it ended up coming down to for me is that these properties
are mutually exclusive. While we can guard against this in a few ways
(target the same backing property, throw an exception), both of these
cases require extra code to be mindful of in some usage where it does
not with a single property, or worse it is silently unexpected behavior.

While it is cleaner to not have to cast, the risk of introducing extra
code complexity for the sake of structure tips the scales.