[Core] Move ObjectPropertyViewModel base to PVM<object>
We don't really need this to be separate, it only helped to diverge
capabilities that arrived in the base value view model but not in the
object one. While it might be a little weird to have Value available for
something that is meant to be edited at a sub-property level, it
wouldn't be the first VM that has a separate "real" value property and
we actually do need a PVM<object> that we can set the value on for
bindings.

This also moves the assignable types request to be fired only when
actually requested.