[Core] Add IPropertyValue
Given that the primary justification for generics was more around the manipulation and handling
of those values, some more generic controls (like previewing values) don't need rigid handling
and instead need a more flexibile option. Given that WPF is already boxing via reflection
it is not unreasonable to provide a way for mac to get the same generic handling without
tons of manual reflection.