Workbooks: Prevent NRE on CALayer.ContentsFormat
It seems that in this branch, using `EnumPropertyViewModel` causes no
editor to be created. `PropertyTableDelegate` does not check for null
editors, resulting in NREs. Add checks, just in case.

In my testing, only `CALayer.ContentsFormat` ends up using
`EnumPropertyViewModel`. For now, comment-out the section in
`PropertiesViewModel.GetViewModel` that creates `EnumPropertyViewModel`,
so that at least a `StringPropertyViewModel` is created.