[mac] Fix editor panel theme handling
DynamicFillBox now allows for colors to be not found, sort of a hack fix
to deal with the fact that we allow PropertyEditorPanel to be used in
xibs, so we can't demand a IHostResourceProvider in the ctor, we have to
accept it later.

This also fixes the fact that the header theming wouldn't update when
the resource provider changes.