[mac] Move default colors to Mac library
Originally we didn't put the colors in the library because VSM integration
would be forced to override them anyway, but it makes sense to have them
as a backup and for any 3rd party consumers, much like the windows theming
has. Also, this means the default setup won't blow up before the overriden
theme provider has been setup.