Fix a11y bug 999307
Windows has the surprising behavior that when an automation group (like a group of combinable checkboxes) has a colon in the name, the group label is not read by the narrator. This causes a problem for the Android designer, which has property names like "app:layout_anchorGravity". We work around this by replacing the colon with a space in the group's automation name.