Use PreviewKeyDown for Ctrl+Space to show menu (#800)
* Use PreviewKeyDown for Ctrl+Space to show menu

Ctrl+Space is the keyboard short that can
be used to show the property editor menu
for a property (with options like Reset,
Convert to Local Value, etc.). Previously, that
was detected via a KeyBinding in XAML, which
worked for text boxes but not controls like
checkbox and button, which handle the
key down event themselves. Now we
use PreviewKeyDown instead, handling that
in C# code, to work for all control
types.

Fixes AB#1491229