Avoid excess margins on left/right of property editor
The property editor (for reasons I don't fully understand,
but maybe due to move away from a Gtk wrapper)
has noticably wider margins on the left & right in 17.x
as compared to 8.x. It's especially noticable on the
right hand side.

This change fixes that, setting the property table style to
NSTableViewStyleFullWidth, so rows fill the full width of
the table.