Fix setting custom values for non-constrained enums
If a predefined property value is set that doesn't match
one of the standard options, then set it as the valueName,
same as if it was passed in the ValueDescriptor. This fixes
showing the custom value in the property editor UI.
Fixes AB#1600201.

Proppy's intention may have been that custom values like
this should be set in the ValueDescriptor instead of Value -
I'm not sure about that. However the Forms XAML code
doesn't do that. Given that, and the fact that there's no harm
is allowing custom values to be specified in the Value instead
of just ValueDescriptor, we now support both.